requirements.txt
streamlit
langchain
langgraph
openai
pandas
scikit-learn
numpy
matplotlib
seaborn
python-dotenv